v1.2
To build a library, we no longer have to use R.require, we can add
config options to the settings of Raccoon to build out the dependencies
of scripts.  Also, all current contributions have been moved to a
contrib directory and the minified RaccoonJS has been removed from the
repo.  There has been extra settings added such as changing the
extension to be different from .js to .min.js, for example, and
changing the delimiter of script names so it can be something like
name>of>script instead of name.of.script.