1.6 Alpha
Many optimizations and a few minor fixes/tweaks.  More to come over the
next few days.  We will see contrib.ui.ScrollPane make use of
contrib.events.Touch as well as a beta framework for page layouts.
This commit contains an overhaul of contrib.utils.Navigator and a
function name change from search to query for
contrib.utils.extensions.HTMLElement.  This search function caused IE 7
and perhaps IE 6 to actually insert the function's toString method
result as a query parameter in the browser's Url if an anchor tag is
clicked.  A simple rename fixed the issue but due note that
HTMLElement.prototype.search is now deprecated.